chore: restructured/renamed time filter directories
